A class for providing a file name reference to a GML document

A class for providing a file name reference to a GML document.

Author

Jussi Jousimo jvj@iki.fi

Super class

rwfs::WFSRequest -> GMLFile

Methods

Public methods

GMLFile$new()

GMLFile$getDataSource()

GMLFile$clone()

Inherited methods

rwfs::WFSRequest$print()

Method new()

Usage

GMLFile$new(srcFile)

Method getDataSource()

Usage

GMLFile$getDataSource()

Method clone()

The objects of this class are cloneable with this method.

Usage

GMLFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.