Downloads response from a WFS and parses the intermediate file

Dispatches a WFS request, saves the response to a file and parses the file. The data can be converted using ogr2ogr of RGDAL. Provides a caching mechanism for subsequent queries on the same data.

See also

Author

Jussi Jousimo jvj@iki.fi

Super class

rwfs::WFSClient -> WFSCachingClient

Methods

Public methods

WFSCachingClient$saveGMLFile()

WFSCachingClient$loadGMLFile()

WFSCachingClient$listLayers()

WFSCachingClient$getLayer()

WFSCachingClient$clone()

Inherited methods

rwfs::WFSClient$getRaster()

rwfs::WFSClient$initialize()

rwfs::WFSClient$setRequest()

Method saveGMLFile()

Usage

WFSCachingClient$saveGMLFile(destFile)

Method loadGMLFile()

Usage

WFSCachingClient$loadGMLFile(fromFile)

Method listLayers()

Usage

WFSCachingClient$listLayers()

Method getLayer()

Usage

WFSCachingClient$getLayer(layer, ...)

Method clone()

The objects of this class are cloneable with this method.

Usage

WFSCachingClient$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.