An abstract class for building a URL reference to a WFS with a caching

The abstract method getURL must be overloaded in a subclass to provide a request URL to a WFS service.

Author

Jussi Jousimo jvj@iki.fi

Super classes

rwfs::WFSRequest -> rwfs::WFSStreamingRequest -> WFSCachingRequest

Methods

Public methods

WFSCachingRequest$getDataSource()

WFSCachingRequest$clone()

Inherited methods

rwfs::WFSRequest$print()

rwfs::WFSStreamingRequest$getCapabilities()

rwfs::WFSStreamingRequest$getFeature()

rwfs::WFSStreamingRequest$getParameters()

rwfs::WFSStreamingRequest$setParameters()

rwfs::WFSStreamingRequest$setPath()

Method getDataSource()

Usage

WFSCachingRequest$getDataSource()

Method clone()

The objects of this class are cloneable with this method.

Usage

WFSCachingRequest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.